refactor(atomic): change base result list for a render function
https://coveord.atlassian.net/browse/KIT-1517
